refactor: move cluster resources to a dedicated cluster directory
